Resolve Github security warning
Github warned me that the Nokogiri version specified in Gemfile.lock had
security issues. This prompted me to update all of the Rspec setup code
to the latest version using sync_spec.

I also drop support for Debian here, because the tests were failing.
Presumably it's trivial to patch it back in, but I don't have any Debian
to test on.